lista completa

+++ Examenes/PP.Usuarios/clases/producto.php
@@ -39,22 +39,25 @@
                     isset($_POST['usuario']) && !empty($_POST['usuario']) &&
                     isset($_FILES['imagen']) && !empty($_FILES['imagen']) ){
                     if( Usuario::existeNombre($_POST['usuario']) ){
-                        //Imagen
-                        $origen = $_FILES["imagen"]["tmp_name"];
-                        $nombreOriginal = $_FILES["imagen"]["name"];
-                        $ext = pathinfo($nombreOriginal, PATHINFO_EXTENSION);
-                        $destinoFoto = "./img/".$_POST['id']."-".$_POST['nombre'].".".$ext;
-                        move_uploaded_file($origen, $destinoFoto);
-                        
-                        $producto = new Producto(   $_POST['id'], 
-                                                    $_POST['nombre'], 
-                                                    $_POST['precio'], 
-                                                    $_POST['cantidad'],
-                                                    $_POST['usuario'],
-                                                    $destinoFoto );
-                        Archivo::guardarUno( './archivos/productos.txt', $producto );
-                        echo 'Se cargo el pruducto: ' . PHP_EOL . $producto->toString();
-
+                        if( !Producto::existeId( $_POST['id'] )){
+                            //Imagen
+                            $origen = $_FILES["imagen"]["tmp_name"];
+                            $nombreOriginal = $_FILES["imagen"]["name"];
+                            $ext = pathinfo($nombreOriginal, PATHINFO_EXTENSION);
+                            $destinoFoto = "./img/".$_POST['id']."-".$_POST['nombre'].".".$ext;
+                            move_uploaded_file($origen, $destinoFoto);
+                            
+                            $producto = new Producto(   $_POST['id'], 
+                                                        $_POST['nombre'], 
+                                                        $_POST['precio'], 
+                                                        $_POST['cantidad'],
+                                                        $_POST['usuario'],
+                                                        $destinoFoto );
+                            Archivo::guardarUno( './archivos/productos.txt', $producto );
+                            echo 'Se cargo el pruducto: ' . PHP_EOL . $producto->toString();
+                        }else{
+                            echo 'Ya existe un pructo con el id: ' . $_POST['id'];
+                        }
                     }else{
                         echo 'No existe un usuario con ese nombre';
                     }
@@ -64,6 +67,35 @@
             }else{
                 echo 'Debe llamarse con el metodo POST';
             }
+        }
+
+        public static function existeId( $id ){
+            $listaDeProductos = Archivo::leerProductos();
+            foreach( $listaDeProductos as $producto ){
+                if( strcasecmp($producto->id, $id) == 0 )
+                    return true;
+            } 
+            return false;
+        }
+
+        /////////////////////////////////////////////////////////////////
+        // LISTAR PRODUCTO
+        /////////////////////////////////////////////////////////////////
+
+        public static function listarProducto(){
+            if( Validaciones::existePeticionGET() ){
+                $productos = Archivo::leerProductos();
+                echo 'LISTA DE PRODUCTOS' . PHP_EOL . PHP_EOL;
+                if( count($productos) > 0){
+                    foreach( $productos as $item ){
+                        echo $item->toString();
+                    }
+                }else{
+                    echo 'No hay Productos para mostrar';
+                }
+            }else{
+                echo 'Debe llamarse con el metodo GET';
+            } 
         }
     }
 ?>
